\name{non.user.level.objects}
\alias{non.user.level.objects}
\alias{[.multic}
\alias{add.heritability}
\alias{all.equal.multic}
\alias{apply.sequential.ids}
\alias{canonical.path.name}
\alias{calculate.coefficients}
\alias{calculate.cor.values}
\alias{calculate.correlations}
\alias{calculate.descriptives}
\alias{calculate.initial.values}
\alias{check.data.for.missing.values}
\alias{check.for.full.data}
\alias{clean}
\alias{collapseMibd}
\alias{combine.null.and.alt.initial.values}
\alias{copy.files.to.tmp.space}
\alias{create.multic.object}
\alias{expand.locus}
\alias{expand.mloci}
\alias{expand.share}
\alias{expandMibd}
\alias{fitted.multic}
\alias{fixibd}
\alias{formula.multic}
\alias{gene.eff}
\alias{get.basename}
\alias{get.cM.from.v.mat}
\alias{get.cm.close.to.peak}
\alias{get.covariance.indices}
\alias{get.df.mix}
\alias{get.dist}
\alias{get.extension}
\alias{get.family.count}
\alias{get.family.sizes}
\alias{get.family.parent.inconsistancies}
\alias{get.major.diagonal.indices}
\alias{get.max.lod.and.locus}
\alias{get.n}
\alias{get.p.mix}
\alias{get.share.order}
\alias{get.top.n.families}
\alias{get.unique.families}
\alias{get.var.covar.labels}
\alias{get.variances.from.file}
\alias{gunzip}
\alias{gzip}
\alias{herit.se}
\alias{is.multic}
\alias{is.parent.offspring}
\alias{is.sibling}
\alias{is.spouse}
\alias{lappend}
\alias{lm.ready.check}
\alias{load.effects}
\alias{load.family.log.likelihoods}
\alias{load.inv.exp.sec.der.fixed}
\alias{load.inv.exp.sec.der.random}
\alias{load.iterations}
\alias{load.residuals}
\alias{load.v.matrices}
\alias{load.v.matrix.file}
\alias{load.var.sandwich}
\alias{load.y.beta.diffs}
\alias{load.y.beta.file}
\alias{make.block.id.pair}
\alias{make.kinship.file}
\alias{max.multic}
\alias{mergeWithKinship}
\alias{missing.value}
\alias{mloci.split}
\alias{mlociCut}
\alias{multic.colnames}
\alias{multic.is.dir}
\alias{multic.kurtosis}
\alias{multic.mkdir}
\alias{multic.rownames}
\alias{multic.skewness}
\alias{multic.strsplit}
\alias{multic.strings.split}
\alias{multic.system}
\alias{multic.unlink}
\alias{normalize.mibd}
\alias{parse.trait.names}
\alias{pchisq.mix}
\alias{plot.family.lods}
\alias{polygene}
\alias{print.summary.multic}
\alias{read.pedindex.out}
\alias{recover.multic}
\alias{remove.file}
\alias{remove.temp.files}
\alias{residuals.multic}
\alias{run.alternative.multic}
\alias{sortIbdCols}
\alias{t.rank}
\alias{trim}
\alias{upp.tri.as.vector}
\alias{using.R}
\alias{validate.parents}
\alias{validate.share.out}
\alias{write.alt.initial.values}
\alias{write.initial.values}
\title{
  This function is not intended for user-level use.
}
\description{
This function was designed and implemented for internal use.  It
is not expected to be used by the user and thus, no significant
documentation exists.
}
% Converted by Sd2Rd version 52193.
