\name{plot.multic}
\alias{plot.multic}
\title{
  Plot a multic object
}
\description{
This is a method for the function plot() for objects inheriting from
class \code{multic}. See plot or plot.default
for the general behavior of this 
function and for the interpretation of x. 
}
\usage{
\method{plot}{multic}(x, ylim=NULL, xlim=NULL, lty=1, col=1, lwd=1, ...)
}
\arguments{
\item{x}{
a \code{multic} object
}
\item{xlim}{
range for x-axis
}
\item{ylim}{
range for y-axis
}
\item{lty}{
line type. Default=1
}
\item{col}{
line color. Default=1
}
\item{lwd}{
line width.  Default=1
}
\item{\dots }{
additional arguments like xlab, ylab, etc.
}
}
% docclass is function
% Converted by Sd2Rd version 52193.
