\name{print.multic}
\alias{print.multic}
\title{
  Use print() on a multic object
}
\description{
This is a method for the function print() for objects inheriting from
class \code{multic}. See print or print.default
for the general behavior of this 
function and for the interpretation of x. 
}
\usage{
\method{print}{multic}(x, ...)
}
\arguments{
\item{x}{
a multic object
}
\item{\dots }{
additional parameters to alter the default behavior of print.multic.
Currently \code{...} only exists to pass 'R CMD
check' tests.
}
}
% docclass is function
% Converted by Sd2Rd version 52193.
