\name{plot.multic}
\alias{plot.multic}
\title{
  Plot a multic object
}
\description{
This is a method for the function plot() for objects inheriting from
class \code{multic}. See plot or plot.default
for the general behavior of this 
function and for the interpretation of x. 
}
\usage{
plot.multic(x, ...)
}
\arguments{
\item{x}{
a \code{multic} object
}
\item{\dots }{
additional arguments like xlim, ylab, etc.
}
}
\keyword{}
% docclass is function
% Converted by Sd2Rd version 1.21.
